pyproject.toml
[build-system]
requires = ["setuptools >= 77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pals_schema"
version = "0.2.0"
dependencies = [
  "pydantic",
  "pyyaml",
  "toml",
]
requires-python = ">=3.10"
authors = [
  { name="Axel Huebl", email="axelhuebl@lbl.gov" },
  { name="Edoardo Zoni", email="ezoni@lbl.gov" },
]
maintainers = [
  { name="Axel Huebl", email="axelhuebl@lbl.gov" },
  { name="Edoardo Zoni", email="ezoni@lbl.gov" },
]
description = "Python implementation for the Particle Accelerator Lattice Standard (PALS)"
readme = "README.md"
license = "BSD-3-Clause-LBNL"
license-files = ["LICEN[CS]E*"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Pydantic :: 2",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development",
]

[project.optional-dependencies]
test = ["pytest"]

[project.urls]
Documentation = "https://pals-project.readthedocs.io"
Repository = "https://github.com/campa-consortium/pals-python"
Issues = "https://github.com/campa-consortium/pals-python/issues"

[tool.setuptools.packages.find]
where = ["src"]
